* beta:
  Update translations from Transifex
  Revert "MBS-13042: Increase default font size (12 -> 14px)"
  Update POT files using the production database
  Update translations from Transifex
  MBS-12308: Clarify the missing artist errors
  MBS-13042: Increase default font size (12 -> 14px)
  MBS-12743: Convert Delete area page to React
  Make $can_delete reusable
  MBS-12851: Hide ratings/tags stats for deleted users (#2810)
  MBS-12740: Convert Delete attribute page to React (#2771)
  MBS-13041: Fix broken Utaten favicon (#2921)
  Make button use the standard font size
  Use named variables for font size over rem
  MBS-12749: Convert cdtoc/attach_filter_artist to React
  Better formatting for InlineSubmitButton
  MBS-12309: Clarify /edit/subscribed also includes collection edits (#2489)
  Specified the OS requirements (#2915)
  MBS-10861: Standardize time before auto-deletion to 48 hours (#2638)
  Update POT files using the production database
  MBS-12769: Convert Delete relationship type page to React (#2768)
  MBS-12768: Convert Delete relationship attribute type page to React (#2767)
  MBS-12774: Convert the Delete release page to React (#2762)
  MBS-12659: Convert user ratings page to React (#2682)
  MBS-12652: Convert entity collection tab to React (#2673)
  MBS-12658: Convert user ratings summary to React (#2679)
  MBS-12779: Treat "n-bit" and "loop" as ETI if in parens (#2773)
  MBS-12785: Wrap things sensibly on release sidebar (#2782)
  MBS-12712: Improve wording for new release adding on disc ID lookup (#2754)
  MBS-12738 / MBS-12739: Add tests
  MBS-12771: Convert Change release quality page to React
  Make Change Quality options translatable
  Add missing space
  MBS-12782: Block adding CritiqueBrainz links
  Remove not actually useful "no results" message
  MBS-12998: Lower the spaghetti level of Quick links
  Actually set isSearch on edit search
  MBS-12989: Add "no vote yet" predicate for subscribed editor refining
  Fix inconsistent use of React$Element
  MBS-12739: Don't assume "single" is ETI unless it's inside parens
  MBS-12738: Don't assume "studio" is ETI unless it's inside parens

+++ lib/MusicBrainz/Server/Controller/Admin/Attributes.pm
@@ -143,10 +143,24 @@ sub delete : Chained('attribute_base') Args(1) RequireAuth(account_admin) Secure
 
         $c->detach;
     }
+
+    $c->stash(
+        component_path  => 'admin/attributes/DeleteAttribute',
+        component_props => {
+            attribute => $attr->TO_JSON,
+            form => $form->TO_JSON,
+        },
+        current_view    => 'Node',
+    );
+
     if ($c->form_posted_and_valid($form)) {
-        $c->model('MB')->with_transaction(sub {
-            $c->model($model)->delete($id);
-        });
+        if ($form->field('cancel')->input) {
+            # Do nothing
+        } else {
+            $c->model('MB')->with_transaction(sub {
+                $c->model($model)->delete($id);
+            });
+        }
 
         $c->response->redirect($c->uri_for('/admin/attributes', $model));
         $c->detach;
